requirements.txt
fastapi==0.135.2
uvicorn==0.34.0
pydantic==2.12.5
google-adk==1.27.5
google-genai==1.9.0
python-docx==1.1.2
pdfplumber==0.11.5
